third_party/blink/web_tests/external/wpt/paint-timing/fcp-only/fcp-out-of-bounds-translate-expected.txt

This is a testharness.js-based test.
FAIL First contentful paint fires due to transform-based intersection with document. assert_equals: First contentful paint marked too early.  expected 0 but got 1
Harness: the test ran to completion.
